change double to decimal in Weight

+++ Luzyce.Core/Models/Lampshade/GetLampshadeNorm.cs
@@ -6,8 +6,8 @@ public class GetLampshadeNorm
     public GetLampshade Lampshade { get; set; } = new GetLampshade();
     public GetVariantResponseDto Variant { get; set; } = new GetVariantResponseDto();
     public int QuantityPerChange { get; set; }
-    public double? WeightBrutto { get; set; }
-    public double? WeightNetto { get; set; }
+    public decimal? WeightBrutto { get; set; }
+    public decimal? WeightNetto { get; set; }
     public string? MethodOfPackaging { get; set; } = string.Empty;
     public int? QuantityPerPack { get; set; }
 }
